Change some file and direcotry names
